Perform span comparisons in inlined representation.
